+++ src/hooks/useCharacter.ts
@@ -7,6 +7,7 @@ export type CharacterStore = {
   updateList: (list: Character[]) => void;
   addList: (item: Character) => void;
   deleteItem: (id: string) => void;
+  importList: (list: Character[]) => void;
 };
 
 export const useCharacterStore = create<CharacterStore>((set) => ({
@@ -16,10 +17,21 @@ export const useCharacterStore = create<CharacterStore>((set) => ({
     localStorage.setItem("characterList", JSON.stringify(list));
     set({ list });
   },
-
   addList: (item) => {
     set((state) => {
-      const list = [...state.list, item];
+      const list = [
+        ...state.list,
+        {
+          id: item.id,
+          icon: item.icon,
+          type: item.type,
+          handle_type: item.handle_type,
+          name: item.name,
+          desc: item.desc,
+          content: item.content,
+          model_config: item.model_config,
+        },
+      ];
       localStorage.setItem("characterList", JSON.stringify(list));
       return { list };
     });
@@ -31,6 +43,29 @@ export const useCharacterStore = create<CharacterStore>((set) => ({
       return { list };
     });
   },
+  importList: (list) => {
+    set((state) => {
+      const newList = state.list.filter((item) => {
+        const find = list.find((i) => i.id === item.id);
+        return !find;
+      });
+
+      const calcList = [...newList, ...list].map((item) => ({
+        id: item.id,
+        icon: item.icon,
+        type: item.type,
+        handle_type: item.handle_type,
+        name: item.name,
+        desc: item.desc,
+        content: item.content,
+        model_config: item.model_config,
+      }));
+
+      localStorage.setItem("characterList", JSON.stringify(calcList));
+
+      return { list: calcList };
+    });
+  },
 }));
 
 export const useCharacterInit = () => {
@@ -38,7 +73,6 @@ export const useCharacterInit = () => {
 
   const init = () => {
     const localCharacterList = localStorage.getItem("characterList");
-
     try {
       updateList(localCharacterList ? JSON.parse(localCharacterList) : []);
     } catch (error) {